various fixes, some formatting changes, etc.
Changed the algorithm for rolling new player stats a little.
Mainly so it would quit stomping on the stack.

Stopped storing player passwords in plain text.  When existing
players log in, their player data file will be auto updated.
This change adds a dependency of libsodium to the project.

Instructions on how to build libsodium can be found at
https://download.libsodium.org/doc/installation/

+++ src/act.movement.c
@@ -420,10 +420,10 @@ ACMD(do_open)
             if (back->keyword) {
               sprintf(buf, "The %s is opened from the other side.", fname(back->keyword));
               /* send_to_room(buf, EXIT(ch, door)->to_room); */
-              act(buf, 0, ch, NULL, (void*) EXIT(ch, door)->to_room, TO_EXIT);
+              act(buf, 0, ch, NULL, (void*) (long) EXIT(ch, door)->to_room, TO_EXIT);
             } else {
               /* send_to_room("The door is opened from the other side.\r\n", EXIT(ch, door)->to_room); */
-              act("The door is opened from the other side.", 0, ch, NULL, (void*) EXIT(ch, door)->to_room, TO_EXIT);
+              act("The door is opened from the other side.", 0, ch, NULL, (void*) (long) EXIT(ch, door)->to_room, TO_EXIT);
             }
           }
         }
@@ -479,10 +479,10 @@ ACMD(do_close)
             if (back->keyword) {
               sprintf(buf, "The %s closes quietly.", fname(back->keyword));
               /* send_to_room(buf, EXIT(ch, door)->to_room); */
-              act(buf, 0, ch, NULL, (void*) EXIT(ch, door)->to_room, TO_EXIT);
+              act(buf, 0, ch, NULL, (void*) (long) EXIT(ch, door)->to_room, TO_EXIT);
             } else {
               /* send_to_room("The door closes quietly.\r\n", EXIT(ch, door)->to_room); */
-              act("The door closes quietly.", 0, ch, NULL, (void*) EXIT(ch, door)->to_room, TO_EXIT);
+              act("The door closes quietly.", 0, ch, NULL, (void*) (long) EXIT(ch, door)->to_room, TO_EXIT);
             }
           }
         }
